+++ R/tables.R
@@ -102,8 +102,8 @@ sp_tables_i <- tibble::tribble(~table_num, ~report_num, ~id,   ~table_code,   ~d
 #' @param year year, numeric, 2015-2018 for some datasets, 2010-2018 for others. Can be a vector of length > 1 (see details).
 #' @param month month, numeric. Must be 3, 6, 9 or 12. Can be a vector of length > 1 (see details).
 #' @param ico ID(s) of org to return, character of length one or more. If unset, returns all orgs. ID not checked for correctness/existence. See <http://monitor.statnipokladna.cz/2019/zdrojova-data/prohlizec-ciselniku/ucjed> to look up ID of any org in the dataset.
-#' @param force_redownload Redownload even if recent file present? Defaults to FALSE.
 #' @param dest_dir character. Directory in which downloaded files will be stored. Defaults to `tempdir()`.
+#' @param redownload Redownload even if recent file present? Defaults to FALSE.
 #'
 #' @return a tibble; see Details for key to the columns
 #' @encoding UTF-8
@@ -178,7 +178,7 @@ get_table <- function(table_id, year = 2018, month = 12, ico = NULL,
   years_months <- expand.grid(y = year, m = month)
   dt_fin <- purrr::map2_dfr(years_months$y, years_months$m,
                             ~get_one_table(dataset_id, .x, .y,
-                                           force_redownload = force_redownload))
+                                           redownload = redownload))
   # onyr <- c(2018) %>% purrr::map_dfr(~ get_table(51101, year = ., month = 12))
 
   }